ai-ml-service/requirements.txt
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.9

# Image Processing & Computer Vision
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
imageio>=2.33.0
imutils>=0.5.4

# Document Analysis & OCR
easyocr>=1.7.1
lxml>=4.9.0
pyzbar>=0.1.9
pyaadhaar>=1.0.0
pytesseract>=0.3.10

# Deep Learning (For OCR & Features)
torch>=2.2.0
torchvision>=0.17.0
deepface>=0.0.86
face-recognition>=1.3.0

# Image Forensics & Math
scikit-image>=0.22.0
scipy>=1.12.0
scikit-learn>=1.4.0
exifread>=3.0.0

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'

# Data Management
pandas>=2.2.0

# Advanced Analytics
xgboost>=2.0.0
joblib>=1.3.2
tqdm>=4.66.0
tf-keras>=2.15.0  # Required for DeepFace background
